ai-chat-analyzer/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
japanize-matplotlib>=1.1.3
wordcloud>=1.9.0
mecab-python3>=1.0.6
unidic-lite>=1.0.8
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
transformers>=4.30.0
torch>=2.0.0
networkx>=3.1
hdbscan>=0.8.30
umap-learn>=0.5.3
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
jupyterlab>=4.0.0
ipykernel>=6.25.0
datasets>=2.14.0
chardet>=5.0.0
chardet>=5.0.0
